+++ test/visual/index.ts
@@ -1,4 +1,3 @@
-import './control_rendering';
 import './freedraw';
 import './generic_rendering';
 import './group_layout';